requirements.txt
Flask==2.3.3
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
Flask-Mail==0.9.1
PyMySQL==1.1.0
